refactor: split EDPA report test into protocol-specific and event group variants
Incorporate InProcessKingdom hmssEnabled/trusTeeEnabled params from #3850,
fix copyright years, fix "retrives" typo, and split
InProcessEdpAggregatorLifeOfAReportTest into DirectOnly, HMSS, TrusTee,
and EventGroup subclasses.

Issue: #3847
