+++ src/main/kotlin/org/wfanet/measurement/integration/common/reporting/v2/BUILD.bazel
@@ -55,7 +55,11 @@ kt_jvm_library(
 kt_jvm_library(
     name = "in_process_edp_aggregator_life_of_a_report_test",
     srcs = [
+        "InProcessEdpAggregatorDirectOnlyReportTest.kt",
+        "InProcessEdpAggregatorHmssReportTest.kt",
         "InProcessEdpAggregatorLifeOfAReportTest.kt",
+        "InProcessEdpAggregatorLifeOfAnEventGroupTest.kt",
+        "InProcessEdpAggregatorTrusTeeReportTest.kt",
     ],
     data = [
         "//src/main/k8s/testing/secretfiles:secret_files",
