+++ src/main/kotlin/org/wfanet/measurement/integration/common/reporting/v2/InProcessMultiEdpReportIntegrationTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Cross-Media Measurement Authors
+ * Copyright 2026 The Cross-Media Measurement Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
