+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorDirectOnlyReportTest.kt
@@ -14,10 +14,8 @@
 
 package org.wfanet.measurement.integration.common.reporting.v2
 
-import kotlinx.coroutines.runBlocking
 import org.junit.ClassRule
 import org.junit.Rule
-import org.junit.Test
 import org.junit.rules.Timeout
 import org.wfanet.measurement.common.db.r2dbc.postgres.testing.PostgresDatabaseProviderRule
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
@@ -29,12 +27,9 @@ import org.wfanet.measurement.integration.deploy.gcloud.SpannerAccessServicesFac
 import org.wfanet.measurement.integration.deploy.gcloud.SpannerDuchyDependencyProviderRule
 import org.wfanet.measurement.reporting.deploy.v2.postgres.testing.Schemata.REPORTING_CHANGELOG_PATH as POSTGRES_REPORTING_CHANGELOG_PATH
 
-/**
- * Implementation of [InProcessEdpAggregatorLifeOfAReportTest] for GCloud backends with Spanner
- * database.
- */
-class GCloudEdpAggregatorLifeOfAReportTest :
-  InProcessEdpAggregatorLifeOfAReportTest(
+/** Implementation of [InProcessEdpAggregatorDirectOnlyReportTest] for GCloud backends. */
+class GCloudEdpAggregatorDirectOnlyReportTest :
+  InProcessEdpAggregatorDirectOnlyReportTest(
     kingdomDataServicesRule = KingdomDataServicesProviderRule(spannerEmulator),
     duchyDependenciesRule = SpannerDuchyDependencyProviderRule(spannerEmulator, ALL_DUCHY_NAMES),
     secureComputationDatabaseAdmin = spannerEmulator,
@@ -53,16 +48,6 @@ class GCloudEdpAggregatorLifeOfAReportTest :
    * TODO(Kotlin/kotlinx.coroutines#3865): Switch back to CoroutinesTimeout when fixed.
    */
   @get:Rule val timeout: Timeout = Timeout.seconds(180)
-
-  @Test
-  fun `HMSS no noise basic report fails when EDP requires Gaussian noise`() = runBlocking {
-    assertHmssReportFailsWhenEdpRequiresGaussianNoise()
-  }
-
-  @Test
-  fun `TrusTee no noise basic report fails when EDP requires Gaussian noise`() = runBlocking {
-    assertTrusTeeReportFailsWhenEdpRequiresGaussianNoise()
-  }
 
   companion object {
     @get:ClassRule @JvmStatic val spannerEmulator = SpannerEmulatorRule()
