+++ src/test/kotlin/org/wfanet/measurement/integration/common/reporting/v2/GCloudEdpAggregatorTrusTeeReportTest.kt
@@ -1,0 +1,67 @@
+// Copyright 2026 The Cross-Media Measurement Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.wfanet.measurement.integration.common.reporting.v2
+
+import kotlinx.coroutines.runBlocking
+import org.junit.ClassRule
+import org.junit.Rule
+import org.junit.Test
+import org.junit.rules.Timeout
+import org.wfanet.measurement.common.db.r2dbc.postgres.testing.PostgresDatabaseProviderRule
+import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
+import org.wfanet.measurement.integration.common.ALL_DUCHY_NAMES
+import org.wfanet.measurement.integration.common.IMPRESSION_QUALIFICATION_FILTER_MAPPING
+import org.wfanet.measurement.integration.deploy.gcloud.InternalReportingServicesProviderRule
+import org.wfanet.measurement.integration.deploy.gcloud.KingdomDataServicesProviderRule
+import org.wfanet.measurement.integration.deploy.gcloud.SpannerAccessServicesFactory
+import org.wfanet.measurement.integration.deploy.gcloud.SpannerDuchyDependencyProviderRule
+import org.wfanet.measurement.reporting.deploy.v2.postgres.testing.Schemata.REPORTING_CHANGELOG_PATH as POSTGRES_REPORTING_CHANGELOG_PATH
+
+/** Implementation of [InProcessEdpAggregatorTrusTeeReportTest] for GCloud backends. */
+class GCloudEdpAggregatorTrusTeeReportTest :
+  InProcessEdpAggregatorTrusTeeReportTest(
+    kingdomDataServicesRule = KingdomDataServicesProviderRule(spannerEmulator),
+    duchyDependenciesRule = SpannerDuchyDependencyProviderRule(spannerEmulator, ALL_DUCHY_NAMES),
+    secureComputationDatabaseAdmin = spannerEmulator,
+    accessServicesFactory = SpannerAccessServicesFactory(spannerEmulator),
+    reportingDataServicesProviderRule =
+      InternalReportingServicesProviderRule(
+        spannerEmulator,
+        reportingPostgresDatabaseProvider,
+        IMPRESSION_QUALIFICATION_FILTER_MAPPING,
+      ),
+  ) {
+
+  /**
+   * Rule to enforce test method timeout.
+   *
+   * TODO(Kotlin/kotlinx.coroutines#3865): Switch back to CoroutinesTimeout when fixed.
+   */
+  @get:Rule val timeout: Timeout = Timeout.seconds(180)
+
+  @Test
+  fun `TrusTee no noise basic report fails when EDP requires Gaussian noise`() = runBlocking {
+    assertTrusTeeReportFailsWhenEdpRequiresGaussianNoise()
+  }
+
+  companion object {
+    @get:ClassRule @JvmStatic val spannerEmulator = SpannerEmulatorRule()
+
+    @get:ClassRule
+    @JvmStatic
+    val reportingPostgresDatabaseProvider =
+      PostgresDatabaseProviderRule(POSTGRES_REPORTING_CHANGELOG_PATH)
+  }
+}
